+++ test/app/module.ts
@@ -1,0 +1,10 @@
+import { Module } from '@nestjs/common';
+
+import { ConsoleModule } from '../../src/module';
+import { ServiceTest } from './service';
+
+@Module({
+    imports: [ConsoleModule],
+    providers: [ServiceTest]
+})
+export class ModuleTest {}